pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doc-quality-gate"
version = "0.1.0"
description = "Implementation Document Quality Gate - review, validate, revise, and score software implementation documents before coding starts"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "pydantic>=2.5.0",
    "typer[all]>=0.9.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.3",
    "rich>=13.7.0",
    "structlog>=24.1.0",
    "litellm>=1.40.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
]

[project.scripts]
dqg = "app.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
per-file-ignores = { "src/app/web/app.py" = ["E501"] }

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
